# api/backend/auth-service/openapi.yml
openapi: 3.1.0
info:
  title: Auth Service
  description: LazyRAG 认证与授权服务（登录、注册、Token、用户/角色/组管理）
  version: 1.0.0
paths:
...
  /api/authservice/group:
    get:
      tags:
      - group
      summary: List Groups
      description: 分页查询用户组列表，支持按关键词、租户筛选
      operationId: list_groups_api_authservice_group_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: tenant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - group
      summary: Create Group
      description: 创建用户组，需填写组名、备注、租户等
      operationId: create_group_api_authservice_group_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/group/{group_id}/basic:
    get:
      tags:
      - group
      summary: Get Group Basic
      description: 查询用户组基础信息（已登录即可，用于服务间回填组名）。
      operationId: get_group_basic_api_authservice_group__group_id__basic_get
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupBasicResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/group/{group_id}:
    get:
      tags:
      - group
      summary: Get Group
      description: 查询指定用户组详情(组名、备注、租户等)
      operationId: get_group_api_authservice_group__group_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - group
      summary: Update Group
      description: 更新指定用户组(组名、备注、租户)；需 user.admin(由网关鉴权)。
      operationId: update_group_api_authservice_group__group_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - group
      summary: Delete Group
      description: 删除指定用户组
      operationId: delete_group_api_authservice_group__group_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/group/{group_id}/permissions:
    get:
      tags:
      - group
      summary: Get Group Permissions
      description: 查询组绑定的权限组；组内成员鉴权时自动拥有这些权限（与角色权限并集）。
      operationId: get_group_permissions_api_authservice_group__group_id__permissions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPermissionsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - group
      summary: Set Group Permissions
      description: 全量设置组的权限组；组内成员自动拥有新权限，无需单独写用户表。
      operationId: set_group_permissions_api_authservice_group__group_id__permissions_put
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupPermissionsBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/group/{group_id}/user:
    get:
      tags:
      - group
      summary: List Group Users
      description: 查询指定用户组内的成员列表(用户 id、用户名、组内角色、租户)
      operationId: list_group_users_api_authservice_group__group_id__user_get
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupUserListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - group
      summary: Add Group Users
      description: 将指定用户批量加入用户组，可指定组内角色(默认 member)，已存在则跳过
      operationId: add_group_users_api_authservice_group__group_id__user_post
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupAddUsersBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/group/{group_id}/user/remove:
    post:
      tags:
      - group
      summary: Remove Group Users
      description: 从用户组中批量移除指定用户
      operationId: remove_group_users_api_authservice_group__group_id__user_remove_post
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRemoveUsersBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/group/{group_id}/user/role:
    patch:
      tags:
      - group
      summary: Set Member Roles Batch
      description: 修改组内成员角色（支持批量）。user_ids 传一个或多个用户 UUID，与 role 一起使用。
      operationId: set_member_roles_batch_api_authservice_group__group_id__user_role_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupMemberRoleBatchBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    GroupAddUsersBody:
      properties:
        user_ids:
          items:
            type: string
          type: array
          title: User Ids
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      type: object
      required:
      - user_ids
      title: GroupAddUsersBody
    GroupBasicResponse:
      properties:
        group_id:
          type: string
          title: Group Id
        group_name:
          type: string
          title: Group Name
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      type: object
      required:
      - group_id
      - group_name
      title: GroupBasicResponse
      description: 用户组基础信息
    GroupCreateBody:
      properties:
        group_name:
          type: string
          title: Group Name
        remark:
          anyOf:
          - type: string
          - type: 'null'
          title: Remark
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      type: object
      required:
      - group_name
      title: GroupCreateBody
    GroupCreateResponse:
      properties:
        group_id:
          type: string
          title: Group Id
      type: object
      required:
      - group_id
      title: GroupCreateResponse
      description: 创建用户组返回
    GroupDetailResponse:
      properties:
        group_id:
          type: string
          title: Group Id
        group_name:
          type: string
          title: Group Name
        remark:
          anyOf:
          - type: string
          - type: 'null'
          title: Remark
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      type: object
      required:
      - group_id
      - group_name
      title: GroupDetailResponse
      description: 用户组详情
    GroupItem:
      properties:
        group_id:
          type: string
          title: Group Id
        group_name:
          type: string
          title: Group Name
        remark:
          anyOf:
          - type: string
          - type: 'null'
          title: Remark
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      type: object
      required:
      - group_id
      - group_name
      title: GroupItem
      description: 用户组列表项
    GroupListResponse:
      properties:
        groups:
          items:
            $ref: '#/components/schemas/GroupItem'
          type: array
          title: Groups
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - groups
      - total
      - page
      - page_size
      title: GroupListResponse
      description: 用户组列表
    GroupMemberRoleBatchBody:
      properties:
        user_ids:
          items:
            type: string
          type: array
          title: User Ids
        role:
          type: string
          title: Role
      type: object
      required:
      - user_ids
      - role
      title: GroupMemberRoleBatchBody
      description: 批量修改组内成员角色，user_ids 支持单个或多个，与 role 一起使用
    GroupPermissionsBody:
      properties:
        permission_groups:
          items:
            type: string
          type: array
          title: Permission Groups
      type: object
      required:
      - permission_groups
      title: GroupPermissionsBody
      description: 组权限全量设置：权限组 code 列表，与角色权限并集后生效，不重复存储
    GroupPermissionsResponse:
      properties:
        permission_groups:
          items:
            type: string
          type: array
          title: Permission Groups
      type: object
      required:
      - permission_groups
      title: GroupPermissionsResponse
      description: 组绑定的权限组 code 列表
    GroupRemoveUsersBody:
      properties:
        user_ids:
          items:
            type: string
          type: array
          title: User Ids
      type: object
      required:
      - user_ids
      title: GroupRemoveUsersBody
    GroupUpdateBody:
      properties:
        group_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Name
        remark:
          anyOf:
          - type: string
          - type: 'null'
          title: Remark
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      type: object
      title: GroupUpdateBody
    GroupUserItem:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        role:
          type: string
          title: Role
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      type: object
      required:
      - user_id
      - username
      - role
      title: GroupUserItem
      description: 组内用户项
    GroupUserListResponse:
      properties:
        users:
          items:
            $ref: '#/components/schemas/GroupUserItem'
          type: array
          title: Users
      type: object
      required:
      - users
      title: GroupUserListResponse
      description: 组内用户列表
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
      description: 通用 ok 返回
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer